Delphi API: add option to load .dll dynamically
If BPSPELLER_STATIC is defined, BPSpeller.dll will be loaded on app startup and will crash the app if loading fails. If it's not defined, an attempt to load library will be taken upon first call to InitializeSpeller function. The downside of this is if InitializeSpeller returns nil, it might be because of lack of memory, unsupported Windows version or DLL link failure.
Also, InitializeSpeller in BPSpellerApi.pas will now validate Windows version and refuse to even attempt to load BPSpeller.dll if it's earlier than 6.2 (Windows 8/Windows Server 2012).

Signed-off-by: Piotr Dałek <git@predictor.org.pl>